[FIX]LINKR/Fixing the problem on edit & delete posts
